+++ Runtime/EconomyPackageInitializer.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Unity.Services.Economy.Internal.Apis.Currencies;
+using Unity.Services.Economy.Internal.Apis.InternalConfiguration;
 using Unity.Services.Economy.Internal.Apis.Inventory;
 using Unity.Services.Economy.Internal.Apis.Purchases;
 using Unity.Services.Economy.Internal.Http;
@@ -9,13 +10,15 @@ using Unity.Services.Authentication.Internal;
 using Unity.Services.Core.Configuration.Internal;
 using Unity.Services.Core.Device.Internal;
 using Unity.Services.Core.Internal;
-using Unity.Services.Economy.Internal.Apis.Configuration;
 using UnityEngine;
 
 namespace Unity.Services.Economy
 {
     class EconomyPackageInitializer : IInitializablePackage
     {
+        const string k_CloudEnvironmentKey = "com.unity.services.core.cloud-environment";
+        const string k_StagingEnvironment = "staging";
+
         [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.BeforeSceneLoad)]
         static void Register()
         {
@@ -24,46 +27,68 @@ namespace Unity.Services.Economy
                 .DependsOn<ICloudProjectId>()
                 .DependsOn<IPlayerId>()
                 .DependsOn<IInstallationId>()
-                .DependsOn<IProjectConfiguration>();
+                .DependsOn<IProjectConfiguration>()
+                .DependsOn<IExternalUserId>();
         }
 
         public Task Initialize(CoreRegistry registry)
         {
+            var projectConfiguration = registry.GetServiceComponent<IProjectConfiguration>();
             var cloudProjectId = registry.GetServiceComponent<ICloudProjectId>();
+            var externalUserId = registry.GetServiceComponent<IExternalUserId>();
             var httpClient = new HttpClient();
 
-            var configuration = new Unity.Services.Economy.Internal.Configuration(null, null, null, GetServiceHeaders(registry));
+            var configuration = new Internal.Configuration(GetHost(projectConfiguration), null, null, null);
+            externalUserId.UserIdChanged += id => UpdateExternalUserId(configuration, id);
 
             IAccessToken accessToken = registry.GetServiceComponent<IAccessToken>();
 
-            IConfigurationApiClient configurationApiClient = new ConfigurationApiClient(httpClient, accessToken, configuration);
+            IInternalConfigurationApiClient configurationApiClient = new InternalConfigurationApiClient(httpClient, accessToken, configuration);
             ICurrenciesApiClient currenciesApiClient = new CurrenciesApiClient(httpClient, accessToken, configuration);
             IInventoryApiClient inventoryApiClient = new InventoryApiClient(httpClient, accessToken, configuration);
             IPurchasesApiClient purchasesApiClient = new PurchasesApiClient(httpClient, accessToken, configuration);
 
-            EconomyService.InitializeEconomy(cloudProjectId, accessToken, registry.GetServiceComponent<IPlayerId>(), configurationApiClient, currenciesApiClient, inventoryApiClient, purchasesApiClient);
+            var analyticsUserId = GetAnalyticsUserId(registry);
+            EconomyService.InitializeEconomy(cloudProjectId, accessToken, registry.GetServiceComponent<IPlayerId>(), configurationApiClient, currenciesApiClient, inventoryApiClient, purchasesApiClient, registry.GetServiceComponent<IInstallationId>().GetOrCreateIdentifier(), analyticsUserId);
 
             return Task.CompletedTask;
         }
 
-        static Dictionary<string, string> GetServiceHeaders(CoreRegistry registry)
+        static string GetAnalyticsUserId(CoreRegistry registry)
         {
-            Dictionary<string, string> headers = new Dictionary<string, string>();
+            var externalUserId = registry.GetServiceComponent<IExternalUserId>();
 
-            string installationId = registry.GetServiceComponent<IInstallationId>().GetOrCreateIdentifier();
-            string analyticsUserId = registry.GetServiceComponent<IProjectConfiguration>().GetString("com.unity.services.core.analytics-user-id");
-
-            // If analytics user id is set, use that, otherwise fallback on the installation id
-            if (!String.IsNullOrEmpty(analyticsUserId))
+            if (!String.IsNullOrEmpty(externalUserId.UserId))
             {
-                headers.Add("analytics-user-id", analyticsUserId);
+                return externalUserId.UserId;
+            }
+
+            return null;
+        }
+
+        static void UpdateExternalUserId(Internal.Configuration configuration, string userId)
+        {
+            if (string.IsNullOrEmpty(userId))
+            {
+                configuration.Headers.Remove("analytics-user-id");
             }
             else
             {
-                headers.Add("unity-installation-id", installationId);
+                configuration.Headers["analytics-user-id"] = userId;
             }
+        }
 
-            return headers;
+        string GetHost(IProjectConfiguration projectConfiguration)
+        {
+            var cloudEnvironment = projectConfiguration?.GetString(k_CloudEnvironmentKey);
+
+            switch (cloudEnvironment)
+            {
+                case k_StagingEnvironment:
+                    return "https://economy-stg.services.api.unity.com";
+                default:
+                    return "https://economy.services.api.unity.com";
+            }
         }
     }
 }